Fix: API와 관계없는 DB 테스트코드 삭제 및 이동
api폴더내에서 api코드만을 담기 위해 기존 db테스트 코드를 홈으로 이동

Resolve: #4
